[ADMIN] Memasukan activity Delete Counter dan Register Counter ke dalam Edit Counter
