+++ app/src/main/res/layout/activity_adm_main.xml
@@ -18,19 +18,6 @@
         <LinearLayout
             android:orientation="horizontal"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-
-            <Button
-                android:layout_width="225dp"
-                android:layout_height="wrap_content"
-                android:text="Registrasi Counter"
-                android:id="@+id/register_button"
-                android:onClick="registerCounter"/>
-        </LinearLayout>
-
-        <LinearLayout
-            android:orientation="horizontal"
-            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:paddingTop="20dp">
 
@@ -40,20 +27,6 @@
                 android:text="Ganti Password Counter"
                 android:id="@+id/update_button"
                 android:onClick="updateCounterPass"/>
-        </LinearLayout>
-
-        <LinearLayout
-            android:orientation="horizontal"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:paddingTop="20dp">
-
-            <Button
-                android:layout_width="225dp"
-                android:layout_height="wrap_content"
-                android:text="Hapus Counter"
-                android:id="@+id/delete_button"
-                android:onClick="deleteCounter"/>
         </LinearLayout>
 
         <LinearLayout
