+++ app/src/main/res/layout/adm_counter_layout.xml
@@ -6,16 +6,18 @@
     <LinearLayout
         android:orientation="horizontal"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="70dp"
 
         android:paddingLeft="10dp"
         android:paddingTop="10dp"
         android:paddingRight="10dp"
-        android:baselineAligned="false">
+        android:baselineAligned="false"
+        android:id="@+id/linearLayout5"
+        android:gravity="center">
 
         <LinearLayout
             android:orientation="vertical"
-            android:layout_width="300dp"
+            android:layout_width="271dp"
             android:layout_height="match_parent"
             android:gravity="left"
             android:paddingLeft="10dp"
@@ -55,35 +57,41 @@
 
         <LinearLayout
             android:orientation="vertical"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:gravity="center|center_vertical">
+            android:layout_width="50dp"
+            android:layout_height="40dp"
+            android:gravity="center|center_vertical"
+            android:paddingRight="10dp">
 
-            <LinearLayout
-                android:orientation="horizontal"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content">
+            <Button
+                android:layout_width="30dp"
+                android:layout_height="30dp"
+                android:id="@+id/edit_button"
+                android:background="@android:drawable/ic_menu_edit" />
 
-                <Button
-                    android:layout_width="25dp"
-                    android:layout_height="25dp"
-                    android:id="@+id/edit_button"
-                    android:background="@android:drawable/ic_menu_edit" />
-            </LinearLayout>
-
-            <LinearLayout
-                android:orientation="horizontal"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:textAppearance="?android:attr/textAppearanceSmall"
-                    android:text="edit"
-                    android:textSize = "10dp"
-                    android:id="@+id/editText" />
-            </LinearLayout>
         </LinearLayout>
+
+        <LinearLayout
+            android:orientation="vertical"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:gravity="center">
+
+            <Button
+                android:layout_width="30dp"
+                android:layout_height="30dp"
+                android:id="@+id/deleteC"
+                android:background="@android:drawable/ic_menu_delete" />
+        </LinearLayout>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:background="@android:drawable/bottom_bar"
+        android:layout_below="@+id/linearLayout5"
+        android:layout_alignParentStart="true"
+        android:paddingLeft="5dp"
+        android:paddingRight="5dp">
     </LinearLayout>
 </RelativeLayout>
